html/browsers/browsing-the-web/navigating-across-documents/011.html

<!doctype html>
<!--/
specs:
  - id: browsing-the-web.html#navigating-across-documents
    steps: [6]

description: |
  1. This test creates an anchor with both an onclick and href.
  2. The test "clicks" the anchor triggering the onclick handler operation, which will attempt to document.write, followed by navigating the iframe to the anchor's `href=""`.

/-->
<title>Link with onclick document.write and href navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<a target="test" onclick="document.write('<script>parent.postMessage(&quot;write&quot;, &quot;*&quot;)</script>');" href="href.html">Test</a>
<script>
var test = async_test();
var messages = [];
test.step(function() {
  document.querySelector("a").click();
});
window.onmessage = test.step_func(function(e) {
  messages.push(e.data);
  if (messages.length === 2) {
    assert_array_equals(messages, ["write", "href"]);
    test.done();
  }
});
</script>
